07/04 changes

+++ Database Example 3/HelloWorldDB/Persons.cs
@@ -1,5 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -9,7 +11,9 @@ namespace Database_Example_3.HelloWorldDB
     public class Persons
     {
 
-        public long PersonID { get; set; }
+        //[Key]
+        //[DatabaseGenerated(DatabaseGeneratedOption.Identity)]
+        public int PersonID { get; set; }
 
 
         public string? FirstName { get; set; }
